some norme
