+++ sources/draw.c
@@ -12,7 +12,17 @@
 
 #include <ft_scop.h>
 
-void	draw(t_env env, t_obj obj, t_mvp matrice)
+static void	init_mvp(t_env env, t_mvp matrice)
+{
+	glUniformMatrix4fv(env.model_id, 1, GL_FALSE,\
+		(const GLfloat*)&matrice.model);
+	glUniformMatrix4fv(env.view_id, 1, GL_FALSE,\
+		(const GLfloat*)&matrice.view);
+	glUniformMatrix4fv(env.projection_id, 1, GL_FALSE,\
+		(const GLfloat*)&matrice.projection);
+}
+
+void		draw(t_env env, t_obj obj, t_mvp matrice)
 {
 	t_state *state;
 
@@ -25,7 +35,8 @@ void	draw(t_env env, t_obj obj, t_mvp matrice)
 	matrice.model = mat_mat(matrice.model, init_identity(state->scale));
 	matrice.model = mat_mat(matrice.model, init_rot_x(state->angle_y));
 	matrice.model = mat_mat(matrice.model, init_rot_y(state->angle_x));
-	matrice.model = mat_mat(init_translation_inv(state->translation), matrice.model);
+	matrice.model = mat_mat(init_translation_inv(state->translation),\
+		matrice.model);
 	init_mvp(env, matrice);
 	glDrawArrays(GL_TRIANGLES, 0, obj.vertices_final.len);
 	glfwSwapBuffers(env.window);
