--- sources/keyboard.c
@@ -1,19 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   keyboard.c                                         :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: iporsenn <marvin@42.fr>                    +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2020/06/19 11:54:31 by iporsenn          #+#    #+#             */
-/*   Updated: 2020/06/19 11:54:32 by iporsenn         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include <ft_scop.h>
-
-void			key_callback(GLFWwindow *win, int key, int scancode, int act,\
-int mods)
-{
-
-}
